✨ Initialize v0.0.1-ideation for master-api iTraders
Project structure is optimized, template files are removed and other necessary imports are fixed in this commit. Project is ready for development.

Imports and configurations are related to v3.0.0 release of template file, and other necessary imports and .env folder is added as per latest/projected development of v3.1.0 release.